+++ packages/wallet-sdk/examples/node-scripts/authHelpers.ts
@@ -1,11 +1,22 @@
 // Copyright ©, 2023-present, Lightspark Group, Inc. - All Rights Reserved
 
-export const getCredentialsFromEnvOrThrow = () => {
+export type EnvCredentials = {
+  accountId: string;
+  jwt: string;
+  pubKey?: string;
+  privKey?: string;
+  baseUrl: string;
+  jwtSigningPrivateKey?: string;
+};
+
+export const getCredentialsFromEnvOrThrow = (): EnvCredentials => {
   const accountId = process.env["LIGHTSPARK_ACCOUNT_ID"];
   const jwt = process.env["LIGHTSPARK_JWT"];
   const pubKey = process.env["LIGHTSPARK_WALLET_PUB_KEY"];
   const privKey = process.env["LIGHTSPARK_WALLET_PRIV_KEY"];
-  const baseUrl = process.env["LIGHTSPARK_EXAMPLE_BASE_URL"] || "api.lightspark.com";
+  const jwtSigningPrivateKey = process.env["LIGHTSPARK_JWT_PRIV_KEY"];
+  const baseUrl =
+    process.env["LIGHTSPARK_EXAMPLE_BASE_URL"] || "api.lightspark.com";
   if (!accountId || !jwt) {
     throw new Error(
       "Missing test credentials. Please set LIGHTSPARK_ACCOUNT_ID and LIGHTSPARK_JWT."
@@ -16,6 +27,7 @@ export const getCredentialsFromEnvOrThrow = () => {
     jwt,
     pubKey,
     privKey,
-    baseUrl
+    baseUrl,
+    jwtSigningPrivateKey,
   };
 };